Fix Base xml test from sql problems on source file
The base xml test was no longer saving a dummy source file before
it saved the dummy sobi fragment.  This caused a data integrity
error when a dummy sobi fragment was saved because the foreign key
field was referencing a source file that did not exist in the
source file table.  This commit brings back the saving of the
source file.