Prevent erroneous new mismatches
Add a first_seen_date_time field to mismatches which represents the date
the mismatch was fist seen. Recurrences of mismatches contain the same
first_seen_date_time as the original. If a mismatch is created after it
has previously been closed the first_seen_date_time is reset to the
current date time to represent a new mismatch.

refs #11091