Consolidate SenateSiteDumpId class
Previously, there were several types of SenateSiteDumpId that id'ed
several different types of dump.  Each of these was reflected in an
implementation of SenateSiteDumpId.  For simplicity's sake, we reduced
this to one single type of dump: the year/session dump.  This commit
simplifies the SenateSiteDumpId class and all of its uses to only handle
this one case.

- Remove SenateSiteDumpSessionId.
- Add dumpTime and year fields to SenateSiteDumpId.
- Modify Json parsing classes to read the "year" field instead of
  "session" to reflect changes in dump code.
- Remove range abstract method.  This was primarily used to get updates
  ranges and to get a reference date time (the end of the date range).
  The updates ranges are no longer used and the reference date time can
  be determined by the dumpTime field.