Adds spotcheck interval reports
These are a refactored version of the weekly reports. Main differences
are they check references YTD instead of over the last week, and they
can run on any schedule (defaults to daily).

Part of this commit fixes a bug in the old weekly calendar report
which caused it to run against unchecked references and therefore never had any
data to check when running. It now correctly checks call calendar
references from the current year.