New test for publish date / action processing
Previously, the BillStatProcessor did not have logic for setting the
publishing date on the first ingestion of a bill. Additionally, the
BillActionParser could not handle 2 \ns in a row (i.e. \n\n).
A new test primarily used with the debugger to make sure the publishing
date and the actions from amendments are processed properly. This test
can also be used to follow or verify any other variable in the
BillStatProcessor.