Make data process run api asynchronous
Currently, the api method to trigger a data process run waits for the
run to complete before a response is returned.  This can take a long
time and can lapse the user's session for longer processor runs.

This commit makes the data processing portion of this method
asynchronous.  The api method now kicks off processing and returns an
unfinished data process log unit.  The id from this unit can be used to
track the run using the data process log api.