Fix/Remove non-mathcing Regex in BillTextUtils
Due to the change in file format, we want to be able to give end users a
choice between HTML enhanced bill text and the format we have been
providing.
This update removes non-matching regex with a java api method.
Additionally, a base bill test, as well as infrustructure to add more
tests easily has been added in BillHTMLparserTest