Fix faulty regex first matches in bill full text
The previous regex used in striping the HTML full text would sometimes
replace 'IN ASSEMBLY' or 'IN SENATE' or in 'SENATE - ASSEMBLY' when they
were not in the header of the bill text.
This new regex will now properly only match them if they are in the
header of the file and no where else.