Fix bill amendment mismatch in spotcheck
The spotcheck for XML-SOBI comparison was giving ~1500 errors for every
bill that was amended. The spot check was using the wrong
getActiveVersion() method. The correct getActiveVersion() has been added
to the BillView class.
Additionally, the JsonOpenelgBillDao has been changed to use methods
instead of being one large method. This dao class is now thread safe.