Add new spotcheck api call specifying a session
Previously the spot check to compare sobi's and xml could not be called
from the browser. The start year would always be 1969.

A new api call has been added that specifies the startyear and ends 2
years after the start year. This can be called from the browser and is
synchronous.