Fix missing amendmentVersions on BillSobiProcessor
When the BillActionAnalyzer was changed to fix the bill actions having
the wrong amendment version, it created a foerign key problem with the
database. A fix was written and was applied to the XmlBillStatProcessor.
Unfortunatly, that was not sufficient, because Sobi files also use the
BillActionAnalyzer and could be affected by this.
The method that fixes this problem has been moved to the
AbstractDataProcessor so it can be accessed by both processors. The
BillSobiProcessor now includes this method in the bill action logic.