Make infrustructure for Openleg Calendar spotcheck
The Openleg Bill spotcheck is not enough to verify that all aspects of
the XML files from LBDC are being processed properly.
Infrustructure for a new spot check has been created and will solve this
problem. This infrustructure includes a new spotcheck enum, a new check
and run service for spotchecks, support to be called from the browser,
and integration with the new JsonOpenlegDaoUtils