Update daybreak senate sponsor formatting
LBDC updated the formatting for senate sponsors in daybreak files
starting July 28, 2017. They have removed the 'CO:' from the sponsor
line which was used to break up the sponsor from co sponsors. Now all
sponsors and cosponsors are in a comma separated list where the first
member is the sponsor and additional members are co sponsors.