Change comparison of bill actions
Previously the checkBillActions method just converted the bill actions
object to a Json string. This would add noise to the comparison.
Now a map is created between the date and the
version of that amendment. It is much clearer and arguably more precise
since the comparison is comparing a targeted peice of data.