Implement interface CalendarEntryList
Currently, we did not have a way of comparing calendars from two
openleg instances. In order to compare them properly, calendars from
both instances need to have a way for identifying them so that they
could be compared apples to apples.
A new interface CalendarEntryList aims to do just that. This interface
was implemented on the CalendarSupEntryView and the ActiveListView. This
ensures that all component types of calendars can be compared with the
same CalendarEntryListId.