Add infrastructure to convert JSON into a calView
The Api of Openleg allows for the extraction of all the data in a
calendar serialized into a Json string. This is done by converting a
calendar view into a Json string.
In order for the Json from another instance of Openleg to be converted
to a calendar view, many method and deafult constructors were made across many classes.
The JsonOpenlegCalendarDao now can properly convert Json into a calendar
view.