Change Calendar Spotcheck ID location
The most recent version of the calendar spotcheck was compaing
ActiveListViews with CalendarSupEntryViews. This was incorrect because
CalendarSupEntryViews are a level below in the inheritance tree.
This has been rectified and now CalendarSupViews and ActiveListViews are
being compared. Additionally, the CalendarSupIdView and the
CalendarActiveListIdView both have a conversion to a
CalendarEntryListId. This allows the objects down the inheritance tree
to call them (i.e. CalendarSupViews and ActiveListViews).
The report service logic has been updated to account for the new
comparison.