Fix index out of bound error
Previously The OpenlegCalendarReportService was trying to match the
reference object to an array list of the same objects from source data.
this would cause index out of bounds errors.
To remedy this problem, the views, CalendarSupView and ActiveListView
now can access their Id field. This allows for creation of the source
object from the CalendarDataService, resolving the index out of bound
issue.