Fix SENCALAL xml files not being parsed
Xml files keep the active list within the <sencalendaractive> xml tag.
Sobi files keep the active list in the same tag, but it is nested within
another tag. <SENATEDATA><sencalendaractive>

The XmlSenCalActiveListProcessor now check the file name extension, and
based off of that extension, provides the XmlHelper with the correct
file path for the active lists. This allows Xml files to be parsed, and
maintains compatibility with Sobi files