Simplify spotcheck of floor calendars
The Openleg to Openleg spotcheck was more complicated than it needed to
be. By adding toString() methods to the CalendarEntryView and to the
CalendarSupEntryView a comparison could be made to a string rather than
to each value seperately. Additionally, spot check mismatch type enums
have been deleted to reflect this simplification.