Rework logic in OpenlegCal report & check service
The previous report logic, while functional had a lot of moving pieces.
Thanks to Sam, the new OpenlegBillReportService is far more compact and
inline with his original vision for this spotcheck report.

Additonally, the OpenlegCalendarCheckService now uses the methods in
BaseSpotCheckService. This makes the check service use an established
check and reduces duplicate code.