OpenlegBillCheckService extends BaseSpotCheckServ
The OpenlegBillCheckService now uses the methods from
BaseSpotCheckService to compare various fields of bills. This shortens
the amount of duplicated code and allows for the reusability of the code
with the BaseSpotCheckServ.

Additionally, The OpenlegBillCheckService methods now take (Content, Reference)
instead of (Reference, Content). This increases clarity and confirms
that the reports aren't shown incorrently in the admin section of
Openleg.