Change Spotcheck UI to reflect position of data
The Spotcheck UI has been updated in two ways.
First, the drop down menu now says Openleg XML - Openleg SOBI.
It used to say Openleg - Openleg which is vague.
Second, when viewing a diff for a mismatch, it now shows XML - SOBI.
it used to say SOBI - XML.
This was inaccurate becase the XML data was really on the left side,
that had the header SOBI.
This became necessary after changing the spotcheck Openleg to Openleg
Spotcheck reports to always showing source data, then reference data.