Add Agenda & Committe id to AgendaCommAddendumView
In order to compare two instances of Openleg Agendas, a common Id is
necessary. CommitteeAgendaAddendumId makes the most sense, as it has all
of the fields to identify a unique addedum.
The prov version of Openleg does not have these fields in the same
object, making it impossible to retrieve the unique id for an addendum once parsed back
from Json. By Addign these fields, that will become possible.