Update SpotcheckReport addObservation method
The Spotcheck report addObservation method was updated to include add
the referenceId and setting the time of the report. All of those actions
happened in code right after adding the observation, so they have been
consolidated into one method. To reflect this change, the
OpenlegBillReportService and the OpenlegCalendarReportService have been
updated. Additionally, that method was stripped from the
JsonOpenlegDaoUtils as this change applied to more spotchecks than the
Openleg to Openleg comparisons