Add Json Deserialization support for AgendaIDs
Since the AgendaCommAddendumView was updated to provide the id's
necessary for comparison, the view structure needed some default
constructors, and in the case of AgendaCommAddendumView, a new field.
These will have to be rolled back one day when everything has all been
merged as it is not necessary to have this extra field.