Reorder spotcheck filter hierarchy
The mismatch filter on the spotcheck report page was designed to work
with the following hierarchy: date->data source->mismatch status->
mismatch type->content type.  This means that the mismatch counts for
each filter element were dependent on the filter selection for all
previous filter elements in the hierarchy.  It doesn't make sense to
have the content type as the last filter because each content type has a
practically unique set of mismatch types.  This commit modifies this
hierarchy so that content type fits between data source and mismatch
status in the hierarchy.

- Refactored summary queries to take new parameters according to the new
  hierarchy.
- Modified mismatch type summary to only contain mismatch types relevant
  to a specific content type.
- Modified layout of mismatch filter so that the content type filter
  comes before the mismatch status and type filters.