Add support for reprint tags
One of the changes for XML 2.0 was to handle reprint tags. Bills get a
new print number when they are sent to the other house, and the other
house wants to change it in some way.

These changes need to be tested still to make sure nothing has been
broken.