Directivize spotcheck table
The tables for each content type on the spotcheck report page have some
minor differences, but most of the code for the three tables is the
same.  Currently each table is specified separately, leading to a lot of
duplicate code.  This commit encapsulates much of the functionality of
the table into a reusable directive, eliminating the duplicate code.

- Removed spotcheck inner controls directive.
- Added spotcheck content tab directive.
- Removed table code from the spotcheck report page, replacing it with
  instances of the content tab directive.