Add button to jump to todays spotcheck report
On the spotcheck report page, it can be difficult to notice that you are
not looking at the report for the current day.  This commit adds a
button to jump to the current day's report.  The button is only active
if you are on a previous day's report, allowing it to serve as a visual
indicator that you are not looking at the most recent data.

This commit also rearranges and relables the top controls on the
spotcheck report page.