Fix spotcheck status summary query
In a previous commit I had modified the  where clause filter in the
existing mismatch section of the mismatch status summary query.  I
changed the logic to filter based on observed date time rather than
first seen date time, because we still want to see mismatches from the
previous session as long as they are still observed this session.  This
was hiding existing issues observed on the same day as the report.

I modified this to use first seen date time again, but left it as a less
than comparison rather than a between to remain in line with my previous
fix.