Init. all published amendments in bill processor
A recent fix to the assignment of amendment versions to actions in the
bill action analyzer has caused new amendments to be referenced in the
actions.  Some of these amendments are not explicitly created before
they are referenced in actions, causing a foreign key constraint
violation when the action is inserted.

This commit modifies the sobi bill processor to ensure the existence of
any bill amendments referenced in actions.