Make BillTextUtils handle regular text
The BillTextUtils class handles the stripping of html from the full text
for a given bill. This works well for text taken from XML files, because
they have html in the body. However, compatibility must be maintained
for sobis. This means that the getStrippedFulltext() method in the
BillAmendment class, used by the BillAmendmentView must play nice with
nonhtml text.