Fix spotcheck mismatch url filter
The filter that generates urls for the openleg content of a spotcheck
mismatch operate based on reference type instead of content type.  These
should be content type so we don't have to write a new filter for every
reference type.

This commit consolidates the existing url generating functions into one
for each content type and makes the contentUrl filter based on content
type.