Disable disabling of diff for spotcheck mismatches
The spotchecks currently disable the diff window for reference/data
missing mismatches.  While this makes sense, we would like to include
some additional information in the diff window (like first and last seen
times) and also would like to be able to use the links to reference and
source content provided in the diff window.

This commit removes any restriction on the spotcheck diff window.