Fix nysenate.gov cal spotcheck issue for sup. cals
The nysenate.gov calendar spotcheck report was generating false positive
supplemental calendar entry errors.  This was because the check service
was checking the entries for ALL openleg supplementals in a given
calendar against a single nysenate.gov supplemental reference.

This commit modifies the supplemental entry check method so that it only
uses openleg supplemental entries from the specific supplemental that is
being checked.