Modify strategy to trigger nysenate.gov spotchecks
The current strategy for running nysenate.gov spotchecks is to attempt
to run a report after each dump fragment is received.  This has the
benefit of running the report immediately after the dump has arrived,
but has issues when there are a lot of dumps coming in at once.  Each
dump is received on a unique thread, and if several threads attempt to
run a report on a fully received dump, all involved threads are locked
down until the report is finished because the method to run the report
is synchronized.

This commit ends the strategy of attempting to run a report after each
request, replacing it with a DataProcessService that checks for full
dumps during the standard data process.