Change method for mapping mismatch keys
The spotcheck dao was parsing the hstore values of keys using a method
that parsed the raw string representation of an hstore.  This method is
somewhat complicated and has the potential to incorrectly map the value
of an hstore.  This occurred for values with commas and spaces (observed
with certain committee names).

This commit modifies the dao to use a simpler approach.  This new
approach uses a built in postgres method to convert an hstore to an
array of alternating keys and values.  The keys and values can be easily
extracted from the array and inserted into a map.