Modify approach for handling lrs outages
In the scraped bill spotcheck, when a bill page is scraped from lrs and
is not parsable (as is the case when lrs is down), it is treated as a
reference missing mismatch.  In this scenario, the bill is not added to
the queue again and these false reference missing mismatches can persist
for a long time.

This commit modifies the approach for handling the lrs outages so that
the scraped bill parser raises a handled exception instead of marking
the bill as missing.  This exception is caught and reported as a warning
notification.  The bill will be added to the scrape queue and will be
scraped until lrs is back online.

Refs #11346