Fix OpenlegAgendaSpotCheck not exec from browser
The OpenlegAgendaReportService was not working as intended from the
browser. When called, it would return a Json Success message instantly
without actually running the report.
Now the report service executes when called from the browser as intended.
The success message that is seen after a call does not appear instantly,
rather it appears after the report is run.