Fix agenda bill check for nysenate.gov spotcheck
The bill/vote check in the nysenate.gov agenda spotcheck didn't
accurately reflect the the way data is stored on nysenate.gov.  I had
assumed that nysenate.gov was only storing the votes, but this was not
true.  Nysenate.gov stores the bills on the agenda info program even if
they aren't voted on.

This commit modifies the check to include all bills on both the agenda
info and vote.  It also renames the mismatch type of this check from
'AGENDA_VOTES' to 'AGENDA_BILLS'.

BE SURE TO RUN THE SQL PATCH!