Fix issue where mismatch view shows wrong diff
The mismatch view has been showing the wrong diff for openleg - lbdc
spotcheck reports.  In this case, the openleg and lbdc diffs have been
swapped.  This is because the diff window for nysenate.gov reports is
laid out so that openleg looks like the reference and nysenate.gov looks
like the observed data and there is no logic to handle this difference
for the diff.

This commit adds better handling for situations where we want to present
openleg as the reference data for a report:

- significantly refactor spotcheck-base.js, separate angular
  declarations from implementation functions.
- Modify reference and content url functions to use openleg as reference
  and nysenate.gov as content for nysenate.gov reports
- Add filters to display the reference and content sources for a
  datasource.
- Add filter that indicates if openleg should be viewed as the
  reference.
- Apply new filters/url functions to the detail window directive.

Fixes #11391