Add order by to bill previous version query
Adds an order by to the direct previous version query in the bill dao so
that the query can handle cases where there are multiple direct previous
versions in the database (this possibility is a relic of the previous
approach for handling previous versions).  With this fix, the query will
select the latest previous version as the direct previous version.