Fix incorrect parsing of actions containing "LOST"
The bill action analyzer interprets any action containing the word
"LOST" to indicate that the bill has lost a floor vote.  This is not
true for a number of actions that contain the word "LOST".

This commit restricts the assignment of a LOST bill status to only
actions with a description exactly matching "LOST".  The commit also
adds a unit test to verify this.