Fix Openleg to Openleg spotcheck HTTP 401 error
All of the openleg to openleg spotcheck tests were failing due to a 401
error. It turned out that some code had been changed in all of the
ReportServices which was caught by @samsto. And had been incorrectly
accessing a property in the app.properties that was meant for something
else.

This is now rectified by creating two new app.properties variables which
are accessed by the environment and the ReportServices