Integrate Opnleg-Opnleg sptchck with new sptchckUI
The new spotcheck from the main dev branch has been merged into the xml
branch, and with that requires a new integration of the spotcheck UI
with that data from those reports.

The Labels have been renamed from( Openleg XML - Openleg Ref ) to ( Openleg Ref - Openleg Source )
As the data is now being displayed in that manner as well.