Create Xml floor vote processor test
The XmlSenFlVotProcessor was the only processor that did not have a unit
test class for it. That has now been corrected with a unit test that
checks that a floor vote xml file can be processed correctly, the result
isn't null, and the correct vote count is recorded from a sample
senflvot file.