Change Section symbol replacement for spotcheck
After doing some research I found that the latin 1 encoding for the
section symbol is Â§ and we are looking for §. This means that somewhere
along the way the text is wrongly encoded to latin 1 from utf-8.

This problem does not affect systems with a default os encoding of
UTF-8. Systems witha default encoding other than UTF-8 I believe are
vulnerable of this issue