Create ProcessYear Obj to govern processing files
In the up coming change of using sobis and xmls reprocessing and current
processing can't stay the same. This new obj will keep track of what
types of files, down to specific parts of a file should be allowed to
process and what should not be allowed to process.

Eventually there will be a class that contians these objects in a map
corresponding with the year it configures for. And an Api that can be
called to change any of these on the fly for ease of use