Create Logic for the ProcessConfig
The ProcessConfig now has a set of logic to be tested and to be improved
upon. The Xml instance is much closer to control over what exactly gets
processed. A set of Api calls still need to be written for accessing
these variales and updating the pend_processing for fragments not deemed
fit to process