Fix alternate session member bug in bill dao
The bill dao is not properly handling updates for bill cosponsors and
multisponsors.  In these update methods, the new cosponsors are compared
to the existing cosponsors.  The method which retrieves existing
cosponsors will replace alternate session members with the primary
session members.  This incorrectly represents the actual session member
ids that are in the database when an existing id is an alternate session
member.

This commit fixes this issue by comparing the new session member ids to
existing session member ids pulled directly from the database with no
substitution.