Fix daybreak publish status check.
The daybreak publish status check did not function correctly in cases
where a bill was restored to a previous amendment.  The check assumes
that all versions prior to the daybreak bill active version are
published.  When an amendment is restored, the latest amendment is still
published, but the active version is returned to a previous amendment.
This causes the check to assume that the former active amendment is not
published.

This commit fixes this issue by using the page file data in the check.
The check still assumes that all amendments prior to the daybreak bill
active amendment are publish, but will add any amendments for which
there are page file entries.

Fixes #11395