Return more member data in member api
The current member api returns only a limited amount of data in the
response if the option "full=true" is present.  There are many other
fields that could be returned including the broken down name and email.

This commit includes these fields by replacing the MemberView returned
in the current response with ExtendedMemberView.