Fix spotcheck missing mismatch labels
The spotcheck mismatch type label filter was performing incorrectly for
nysenate.gov spotchecks, in which we display OpenLeg as if it were the
reference.  Observed data missing mismatches were being displayed as
nysenate.gov missing mismatches.

I have reworked the logic in the mismatch type lable filter so that it
switches reference/observed data missing mismatches if the data source
dictates it.

Refs #11587