Allow resolution of nysenate.gov pub status errors
NYSenate.gov does not yet have an publish status field for bill
amendments, so any unpublished bills should not be present at all.  This
presents a problem if we are trying to resolve a prior spotcheck
mismatch where NYSenate.gov has a bill that has been unpublished because
if the bill is now correctly not present, it will not be checked and
therefore not resolved.

This commit adds some new code in the method that handles reference
missing errors.  This new code separates bills missing from NYSenate.gov
into to categories, one for bills published on openleg, and one for
unpublished bills.  The published bills are given ref missing
observations as usual, but the unpublished bills that are not present in
the dump get empty observations, indicating that they were checked and
no errors found.

Fixes #11587