Optimize OpenLeg Bill Spotcheck report
The Openleg bill spotcheck report has been changed to get some of the
bills and check them and then repeat, as opposed to grabbing all of the
bills and then checking them all. This allows the logs to show progress
of the check and I've sped up the time it takes for completing the test.
I am seeing about 10-20 sec's improvment which im sure will translate to
a greater benefit on the the erver with more resources