Properly set email field for members
The email field was not being set when members were being retrieved from
the database.  It was also not being set in the member view constructor,
resulting in all emails being null.

This commit adds code to query and set the email in the member dao and
also properly set the email field in the ExtendedMemberView constructor.