Fix member query
In a prior commit, I added the email field to the member query row
mapper and to the select statement.  I didn't realize that there was
another select statement using the same row mapper.  That select
statement did not include the email field, so there was an error every
time the row mapper attempted to reda the email field.

This commit adds the email field to the second select statement.