Modify update range logic for agenda api
The agenda updates api takes a closed open datetime range as its
parameter.  This is problematic in practice because the typical use case
involves passing the last received update datetime as the "from"
parameter.  This means that you will receive that same update again
because its update time is >= the datetime in the "from" parameter.  The
other updates apis use an open range.

This commit changes the date range parsing logic to query for an open
range based on the input "from" and "to" datetimes.  Also, the
@RequestMapping annotations have been modified to prevent sub-second
precision times from being lopped off due spring thinking they are file
extensions.