Modifies the Nyseante agenda spotcheck to better compare votes
Previously we compared the nysenate addendum votes the corresponding
openleg vote addendum. This is innacurate because the bills on an
openleg vote addendum do not need to be the same bills which are on the
info addendum.

This change compares bills on the nyseante addendum to the openleg info
addendum. To compare the vote data we seach through the all vote
addendums to find the votes for a bill.

fixes #11648