Fix "Date Reported" in spotcheck detail window
The "Date Reported" field in the spotcheck detail window shows the
current report date, which is useless.  This commit modifies that field
to display the date that the mismatch was last seen.  The label is also
changed to "Last Reported" to reflect that it is a date and a time.

Fixes #11649