Remove tests concerning xml collate and ingest
These tests have been removed because on the dev server when mvn verify
is run, all files in the incoming are moved over to the archive without
ever being processed.This causes huge problems for the spotcheck
reports. Additionally, at this point we know the xml project can handle
the real live runtime of these events. These tests no longer add value
to the project