Refactor spotcheck mismatch diff info card layout.
Modify the content id section of the spotcheck mismatch diff view
window.  Certain elements in this section are duplicated for each
content type.  This commit refactors this section so that it is not
duplicated for each content type.  Instead there are unique fields for
each content type in the same panel.  Also modify the styling of the
panel.

Fixes #11649