Allow elasticsearch connection retries.
There have been issues on the server during automatic redeployments
where elasticsearch is not fully initialized when Openleg is starting
up, causing the openleg startup to fail.  Elasticsearch is always
started before Openleg, but there are components of es startup that are
asynchronous and may not be ready for openleg.

This commit alters the behavior of the elasticsearch connection process
to allow a certain number of retries if the connection is a failure.
This number is configurable in app.properties.