Implement safer file writes in sensite dump dao.
There has been a recurring problem with the nysenate.gov spotcheck dumps
where it attempts to read a dump file while its being written.  This
occurs because the dumps are saved asynchronously on different threads
from the data processor.

This commit attempts to avoid this issue by saving dumps to a temporary
file, which is renamed to the desired dump file name after it is fully
saved.

Fixes #11694