Add pipelining utility
Adds a utility for creating a task pipeline where a sequence of tasks
can be run in parallel.  The pipeline is constructed by adding
individual tasks to a pipeline builder using functional interfaces.
Then, new threads are created to run each task.  Objects are fed into
the pipeline and passed between tasks of the pipeline using thread safe
blocking queues.  The pipeline is finished when all nodes are finished.
Each node is finished when their queue is empty and the previous node is
finished.  The result of the pipeline can be retrieved using a
CompletableFuture.