Fix previous task registry in pipeline builder
There was a bug in the pipeline builder where the task for the previous
node was used for all instances of a task with parallel instances.  This
could cause the pipeline to complete before one of the parallel
instances was done.

This commit links parallel tasks in a chain together, so that there is a
linked sequence of tasks from start to finish that leaves out no task.