Improve co/multisponsor check in daybreak report.
Improves the formatting of the co and multi sponsor checks in the
daybreak report.  These checks would take the shortnames of all
co/multisponsors and print them out space separated on a single line.
This was difficult to read, as it required a great deal of side to side
scrolling.

This commit modifies this check so that the checked sponsor list is
newline separated.  Also, the sponsors are sorted in alphabetical order.