Fix blocking issues with pipeline
There have been blocking issues with the new Pipeline system.  The
blocks occur when a consumer task is repeatedly trying to read from a
queue, rendering a producer unable to push to the queue.

This is fixed by using the poll method instead of the non-concurrent
standard isEmpty check.  The poll method puts the calling thread to
sleep until an item is available in the queue, removing the need to
constantly check and block access to the queue.