Remove nano seconds in publish date time comp.
Xml files offer us a lot of data that we could not get from sobis. One
peice of that data is from the file name itself. Xml files allow us to
use nano second precision whereas sobis we can only specify to the
second.
We have decided to maintian the extra precision in our data and I have
changed the BillCheckService to disregard the nanoseconds when comparing
the published date time. This should resolve thousands of errors on the
OL - NYS GOV spotcheck report