Revert processconfig change for 2017
For a currently unknown reason the process config does not allow any
more fragments to process after reaching about 18000. While calendars
and agendas and active lists are disabled in xml for 2017 this error
occurs.
I plan on revisiting this issue at some point in the near future,
however right not the xml env needs to be rebuilt which is of a much
higher priority