Fix spotcheck report save log statement.
Fixing an error in a recently modified log statement for when a
spotcheck report is saved.  I thought that setting ignored = true in the
report mismatch count would return a count that includes ignored
mismatches.  This actually returns a count of only ignored mismatches.
Fixed to include both counts in the log.