Set elastic index settings on index create.
There is one instance where we use a custom elasticsearch index setting.
We have been setting this manually using the REST api, but the setting
is cleared whenever the index is rebuilt.

This commit adds the ability to configure custom settings to be set on
index creation.  By default, all indices use the default settings, but
the settings method can be overridden to use custom settings.  The
method has been overridden for the one custom setting that we use:
max_result_window in the bill index.