Initialize new amendments with current law code and law section
Often, when a bill is amended, we are getting Law Code and Summary data
in the daybreak files before we get it in the SOBI/XML files. This makes
us believe new amendments should be initialized with the previous active
amendments law info.

This should also eliminate the LBDC-Openleg Law/Summary mismatch false
positives.

refs #11740