Fix Rules Committee from COM error
S6745 is a bill from the rules committee, however, it was appearing as
from the Rules Committee via COM which is incorrect. I have adjusted the
if statement controlling the aditional rules sponsor and resolved the
issue. Additionally, my change does not break A40001 which is supposed
to have a sponsor with the rules committee