Change the way spotcheck notifications work.
We want to get notified every time a spotcheck is run, not just when
there are errors.  This commit modifies the existing system, removing
the "all clear" spotcheck notification, and making so that the
notifications for each spotcheck report are always posted, no matter how
many mismatches are generated.

Fixes #11820