Change Process Config for new prod db test
The process config has been changed for the new db test to simulate how
it will work on prod. I believe for 2017 we only want the bill text from
html, and in 2018 while it will start out as sobi data, we will
reprocess the data with xml files.
 After a full reprocess of 2017-18 data we will have a mixed 2017 year
and a full xml year