Change member ES index document structure.
Modifies the documents that are indexed into the member elasticsearch
index.  Instead of adding a single document for each member for each
session, we add one document for each person for each house.  The
document is now the json from the FullMemberView, which contains records
for each session the member was active.