Handle E.S. limit params above max result window.
Elasticsearch will not accept limit parameters that are higher than an
index's max result window.  This causes issues when limit is set to ALL
which resolves to the java max int value.

This commit reduces limit parameters that exceed the max result window
to the m.r.w value.  This should not cause any additional problems,
since the m.r.w. value is generally much greater than the expected
number of docs in an index.