Simplify process config
The process config for 2017 wasn't properly processing the fragments. It
would only allow one fragment to process at a time which was unintended.
I have simplified the process config's overarching variables which seems
to allow it all to process properly locally. I want to see how this
performs on the server. If it works, then I will clean up my comments
for a prod release