Tweak bill session api to address import issues.
There have been some minor issues with the bills for session api, which
is commonly used for importing bill data.  This api is typically called
multiple times with pagination, since there are ~20k bills per session.
The problem arises when there are bill updates during the import
process, causing the order of the bill query to change.  With a
different order, some bills may appear again in further result windows,
and some bills may never appear due to being shifted behind the current
window.

This is addressed by setting the default order of the results to the
published date time ascending.  Published date time is only set when a
bill is first saved to openleg, so the order should not change in the
middle of a query.

Additionally, I added an "idsOnly" parameter enabling the user to
get the results as simple bill ids.