Modify AgendaCommView to include all votes.
The agenda json was not providing all of the votes in some
circumstances.  This was happening because the AgendaCommView ignored
votes in addenda that did not contain meeting info.  We recently
discovered that vote and info addenda are independent of each other, so
these votes still need to be included.

To fix this, I added logic to include votes for addenda that do not have
meeting info.

Fixes #11868